+++ CHANGELOG.md
@@ -2,6 +2,13 @@ CHANGELOG
 =========
 
 
+2.0.0 (2021-XX-XX)
+------------------
+
+- Use `symfony/config` to process `theme.yml` manifests.  
+The YAML syntax changed.  
+**Upgrade path:** The `layouts` key moved one level up in the `theme.yml` syntax
+
 1.0.0 (2021-03-24)
 ------------------